backend/requirements.txt
arrow==1.2.3
asgiref==3.7.2
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.2
django-picklefield==3.1
django-postgres-copy==2.7.3
django-q==1.3.9
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.12.3
gdown==4.7.1
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
install==1.3.5
joblib==1.3.2
nltk==3.8.1
numpy==1.25.2
oauth2client==4.1.3
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
protobuf==4.24.3
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDrive2==1.17.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==3.5.3
regex==2023.8.8
requests==2.31.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
tomotopy==0.12.5
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
